rodan-main/code/requirements.txt
backports.ssl-match-hostname==3.4.0.2
celery==4.4.7
certifi==14.05.14
django-cors-headers==3.0.2
django-decorator-include==1.3
django-environ==0.4.5
django-extensions==1.6.7
django-filter==1.1.0
django-guardian==1.4.5
django-health-check==2.4.0
django-sortedm2m==1.1.1
django-websocket-redis==0.5.2
Django==1.11.29
djoser==1.5.1
djangorestframework==3.6.4
flower==0.9.3
gevent==1.3.6
gnureadline==6.3.3
gunicorn==19.10.0
jsonfield==1.0.0
jsonschema==2.5.1
lxml==4.6.3
Markdown==2.5.1
mimeparse==0.1.3
model-mommy==1.6.0
pathlib2==2.3.6
Pillow==6.2.2; python_version < "3.4"
protobuf==3.17.3
psycopg2==2.8.4
pybagit==1.5.0; python_version < '3.0'
-e git+https://github.com/deepio/pybagit.git@4eea813148bc590896000f57cde35fb86443d2f4#egg=pybagit ; python_version >= '3.6'
pyparsing==2.0.3
pypng==0.0.18
python-dateutil==2.5
python-magic==0.4.22
PyYAML==5.4
readline==6.2.4.1
redis==2.10.3
uWSGI==2.0.18
Werkzeug==1.0.1

# pil_rodan
Pillow==8.2.0; python_version > "3.4"

# interactive_classifier

# pixel_wrapper
numpy==1.16.6; python_version < "3.4"
opencv-python==4.2.0.32; python_version < "3.4"

# text_alignment
#numpy==1.17.3; python_version > "3.4"
#tensorflow==2.3.2; python_version > "3.4"
#scikit-image==0.15.0; python_version > "3.4"
#calamari-ocr==1.0.5; python_version > "3.4"
#Unidecode==1.0.22; python_version > "3.4"
